Refactor Pinecone integration: rename app to 'pinecone-assistant', remove deprecated search functionality, and add new assistant context handling.
