+++ pinecone-assistant/actions/search.ts
@@ -20,15 +20,8 @@ interface QueryResult {
 }
 
 /**
- * @title Get Assistant Context
- * @description Retrieves relevant document snippets from the assistant's knowledge base.
- * Returns an array of text snippets from the most relevant documents.
- * The snippets are formatted as JSON objects with the fields:
- * - file_name: The name of the file containing the snippet
- * - pages: The pages of the file containing the snippet
- * - content: The snippet content
- * You can use the 'top_k' parameter to control result count (default: 15).
- * Recommended top_k: a few (5-8) for simple/narrow queries, 10-20 for complex/broad topics.
+ * @title Get Context
+ * @description Retrieves relevant document snippets from the assistant's knowledge base.Returns an array of text snippets from the most relevant documents. The snippets are formatted as JSON objects with the fields: - file_name: The name of the file containing the snippet - pages: The pages of the file containing the snippet - content: The snippet content You can use the 'top_k' parameter to control result count (default: 15). Recommended top_k: a few (5-8) for simple/narrow queries, 10-20 for complex/broad topics.
  */
 const action = async (
   props: Props,
@@ -41,7 +34,7 @@ const action = async (
     }
 
     const response = await ctx.client
-      ["POST /chat/:assistant_name/context"]({
+      ["POST /assistant/chat/:assistant_name/context"]({
         assistant_name: ctx.assistant,
       }, {
         body: {
