+++ pinecone-assistant/utils/client.ts
@@ -1,0 +1,11 @@
+import {
+  ChatContextRequest,
+  ChatContextResponse,
+} from "./types.ts";
+
+export interface PineconeAPI {
+  "POST /assistant/chat/:assistant_name/context": {
+    response: ChatContextResponse;
+    body: ChatContextRequest;
+  };
+}
