--- pinecone/utils/client.ts
@@ -1,18 +1,0 @@
-import {
-  ChatContextRequest,
-  ChatContextResponse,
-  SearchRequest,
-  SearchResponse,
-} from "./types.ts";
-
-export interface PineconeAPI {
-  "POST /records/namespaces/:namespace/search": {
-    response: SearchResponse;
-    body: SearchRequest;
-  };
-
-  "POST /chat/:assistant_name/context": {
-    response: ChatContextResponse;
-    body: ChatContextRequest;
-  };
-}
